+++ Source/Entities/Main/UserCompany.cs
@@ -5,11 +5,11 @@ namespace MultiSoftSRB.Entities.Main;
 public class UserCompany
 {
     public long UserId { get; set; }
-    public long CompanyId { get; set; }
+    public short CompanyId { get; set; }
     public bool IsPrimary { get; set; }
     public AccessType AccessType { get; set; } = AccessType.Direct;
     
-    public long? AgencyCompanyId { get; set; }
+    public short? AgencyCompanyId { get; set; }
         
     // Navigation properties
     public virtual User User { get; set; } = null!;
