+++ Source/Features/Admin/Consultants/AssignCompany/Request.cs
@@ -3,5 +3,5 @@ namespace MultiSoftSRB.Features.Admin.Consultants.AssignCompany;
 sealed class Request
 {
     public long ConsultantUserId { get; set; }
-    public long CompanyId { get; set; }
+    public short CompanyId { get; set; }
 }
