+++ Source/Features/Auth/SwitchCompany/Request.cs
@@ -2,5 +2,5 @@ namespace MultiSoftSRB.Features.Auth.SwitchCompany;
 
 sealed class Request
 {
-    public long CompanyId { get; set; }
+    public short CompanyId { get; set; }
 }
